CLI-1066 Move HTTP status constants next to the server client
